Restructure directories & remove `src`
