+++ index.html
@@ -15,5 +15,5 @@
     </main>
   </body>
 
-  <script type="module" src="/src/web.js"></script>
+  <script type="module" src="/web.js"></script>
 </html>
